Move more functionality into presenter
All ContentResolver work in MainActivity has now been moved to
Presenter. DialogFragments have been moved back to MainActivity; they
now forward events to the Presenter when appropriate.
