--- GreekReference/src/main/java/com/benlinskey/greekreference/dialogFragments/ClearLexiconFavoritesDialogFragment.java
@@ -1,62 +1,0 @@
-/*
- * Copyright 2015 Benjamin Linskey
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.benlinskey.greekreference.dialogFragments;
-
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.app.DialogFragment;
-import android.content.ContentResolver;
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.widget.Toast;
-
-import com.benlinskey.greekreference.R;
-import com.benlinskey.greekreference.data.appdata.AppDataContract;
-
-/**
- * A {@link DialogFragment} that asks the user to confirm that he or she wishes to clear the
- * lexicon favorites list. If the user answers in the affirmative, the list is cleared.
- * Otherwise, the dialog is dismissed and no further action is taken.
- */
-public class ClearLexiconFavoritesDialogFragment extends DialogFragment {
-    @Override
-    public Dialog onCreateDialog(Bundle savedInstanceState) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        builder.setMessage(R.string.clear_lexicon_favorites_dialog_message);
-
-        builder.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                ContentResolver resolver = getActivity().getContentResolver();
-                resolver.delete(AppDataContract.LexiconFavorites.CONTENT_URI, null, null);
-
-                String msg = getString(R.string.toast_clear_lexicon_favorites);
-                Toast toast = Toast.makeText(getActivity(), msg, Toast.LENGTH_SHORT);
-                toast.show();
-            }
-        });
-
-        builder.setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                dialogInterface.dismiss();
-            }
-        });
-
-        return builder.create();
-    }
-}
